PostgreSQL.Embedding/PostgreSQL.Embedding/Assets/Scripts/requirements.txt
modelscope
FlagEmbedding
numpy==2.0.2
python-dotenv